+++ src/pj_gridinfo.c
@@ -227,7 +227,7 @@ int pj_gridinfo_load( projCtx ctx, PJ_GRIDINFO *gi )
 
             if( pj_ctx_fread( ctx, row_buf,
                               sizeof(double), gi->ct->lim.lam * 2, fid )
-                != 2 * gi->ct->lim.lam )
+                != (size_t)( 2 * gi->ct->lim.lam ) )
             {
                 pj_dalloc( row_buf );
                 pj_dalloc( ct_tmp.cvs );
@@ -305,7 +305,7 @@ int pj_gridinfo_load( projCtx ctx, PJ_GRIDINFO *gi )
 
             if( pj_ctx_fread( ctx, row_buf, sizeof(float),
                               gi->ct->lim.lam*4, fid )
-                != 4 * gi->ct->lim.lam )
+                != (size_t)( 4 * gi->ct->lim.lam ) )
             {
                 pj_dalloc( row_buf );
                 pj_dalloc( ct_tmp.cvs );
@@ -370,7 +370,7 @@ int pj_gridinfo_load( projCtx ctx, PJ_GRIDINFO *gi )
         }
 
         if( pj_ctx_fread( ctx, ct_tmp.cvs, sizeof(float), words, fid )
-            != words )
+            != (size_t)words )
         {
             pj_dalloc( ct_tmp.cvs );
             pj_release_lock();
